upgrade 05-09-2026, v.3.0 Profiles

+++ CC2/Source/allegext.cpp
@@ -3785,7 +3785,7 @@ void trim_trailing_new_row(char *str) {
 		return; // Handle null or empty strings
 	}
 
-	int i = strlen(str) - 1; // Start from the end of the string
+	int i = (int)strlen(str) - 1; // Start from the end of the string
 
 	// Iterate backwards, identifying trailing new row characters
 	while (i >= 0 && ((unsigned char)str[i])=='\n') {
